[dx11] remove old files and add DSV creation & index buffers
Removes old files leftover in the backend directory.

Adds depth stencil views and binding them to the pipeline. Also adds
index buffers to be bound and some more quick & dirty "bind descriptor
set to stage & binding" for buffers.
